Added disconnect events for native
Simply run the rcv_disconnect_events_native function to view disconnect
events that happen on the native version. As to when it happens on the
naia part, this library simply doesn't support that (yet).

The naia portion of this could be done by constantly watching the 
naia.connections part, and reporting and storing any removals in a Vec.